requirements.txt
asgiref==3.7.2
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
ConfigParser==6.0.0
contextlib2==21.6.0
cryptography==41.0.7
Cython==3.0.6
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
ipython==8.18.1
ipywidgets==8.1.1
jnius==1.1.0
keyring==24.3.0
lockfile==0.12.2
mock==5.1.0
mysql_connector_repackaged==0.3.1
numpy==1.26.2
passlib==1.7.4
paypalrestsdk==1.13.3
Pillow==10.1.0
protobuf==4.25.1
pyOpenSSL==23.3.0
python-dotenv==1.0.0
python_bcrypt==0.3.2
railroad==0.5.0
simplejson==3.19.2
Sphinx==7.2.6
thread==0.1.2
tornado==6.4
trove_classifiers==2023.11.29
truststore==0.8.0
urllib3_secure_extra==0.1.0
watchdog==3.0.0
xmlrpclib==1.0.1
zstandard==0.22.0
